server/requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.1.0
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.4
